refactor: change login and register page architecture
