+++ src/components/Canvas/useUploadImage.ts
@@ -2,7 +2,7 @@ import { addDoc, collection, doc, updateDoc } from 'firebase/firestore';
 import { getDownloadURL, ref, uploadString } from 'firebase/storage';
 import { useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { storage, db } from '../../services/firebase/config';
+import { storage, db } from '../../api/config';
 import { User } from 'firebase/auth';
 import { ImageListItemType } from '../ImageGallery/ImageGallery';
