+++ src/components/ImageGallery/ImageGallery.tsx
@@ -5,7 +5,7 @@ import { User } from 'firebase/auth';
 import { useAppDispatch } from '../../store/store';
 import { setEditImageData } from '../../store/editImageSlice';
 import { useNavigate } from 'react-router-dom';
-import { EDITOR_ROUTE } from '../../utils/constants/routes';
+import { EDITOR_ROUTE } from '../../constants/routes';
 import { deleteDocument } from '../../services/firebase/Documets.service';
 import ImageItem from './ImageItem/ImageItem';
 import { useState } from 'react';
