+++ src/components/ToolsPanel/ToolsPanel.tsx
@@ -15,10 +15,10 @@ import BrushIcon from '@mui/icons-material/Brush';
 import PanoramaFishEyeIcon from '@mui/icons-material/PanoramaFishEye';
 import Crop54Icon from '@mui/icons-material/Crop54';
 import HorizontalRuleIcon from '@mui/icons-material/HorizontalRule';
-import { HOME_ROUTE } from '../../utils/constants/routes';
+import { HOME_ROUTE } from '../../constants/routes';
 import { useAppDispatch, useAppSelector } from '../../store/store';
 import { setToolValue } from '../../store/toolSlice';
-import { Tools } from '../../utils/constants/tools';
+import { Tools } from '../../constants/tools';
 import { setToolColorValue } from '../../store/toolColorSlice';
 import { HexColorPicker } from 'react-colorful';
 import { setPenSizeValue } from '../../store/penSizeSlice';
