+++ src/modules/LoginForm/constants/constants.ts
@@ -1,0 +1,6 @@
+export const formInfo = {
+  title: 'Login',
+  Subtitle: 'Don`t have an account?',
+  navigateRoute: '/reg',
+  navigateLinkText: 'SIGN UP',
+};
