+++ src/modules/RegistrationForm/constants/constants.ts
@@ -1,0 +1,6 @@
+export const formInfo = {
+  title: 'Signup',
+  Subtitle: 'Already have an account?',
+  navigateRoute: '/login',
+  navigateLinkText: 'LOGIN',
+};
