+++ src/modules/RegistrationForm/index.ts
@@ -1,0 +1,1 @@
+export { RegistrationForm } from './components/RegistrationForm';
