+++ src/pages/EditorPage/EditorPage.tsx
@@ -1,7 +1,7 @@
 import { useAuthState } from 'react-firebase-hooks/auth';
 import Canvas from '../../components/Canvas/Canvas';
 import ToolsPanel from '../../components/ToolsPanel/ToolsPanel';
-import { auth } from '../../services/firebase/config';
+import { auth } from '../../api/config';
 
 function EditorPage() {
   const [user] = useAuthState(auth);
